+++ GGT/modules/GMTL/Test/TestSuite/TestCases/QuatCompareTest.cpp
@@ -8,8 +8,8 @@
  *
  * -----------------------------------------------------------------
  * File:          $RCSfile: QuatCompareTest.cpp,v $
- * Date modified: $Date: 2003-02-06 01:12:27 $
- * Version:       $Revision: 1.2 $
+ * Date modified: $Date: 2004-09-16 19:40:33 $
+ * Version:       $Revision: 1.3 $
  * -----------------------------------------------------------------
  *
  *********************************************************** ggt-head end */
@@ -307,7 +307,7 @@ namespace gmtlTest
             ++true_count;
          if (gmtl::isEqual( src_quat44,  test_quat44, 0.8f ))
             ++true_count;
-         if (gmtl::isEqual( src_quat101, test_quat101, 111.1 ))
+         if (gmtl::isEqual( src_quat101, test_quat101, double(111.1) ))
             ++true_count;
       }
